Adding more design flair to deal rows.
